Refactor supplier deletion flow: Replace `DeleteSupplierBtn` with `DeleteSupplierModal` for better user experience and confirmation handling. Integrate modal into `SupplierDetails` and `SuppliersGrid` components. Improve deletion logic, add loading states, and enhance toast notifications for feedback.
